requirements.txt
pyTelegramBotAPI
python-dotenv
python-telegram-bot
matplotlib
apscheduler
pytz